+++ samples/Tutorial/Set.fs
@@ -38,13 +38,13 @@ let interpret add remove (state : Fold.State) =
         [   if adds.Length <> 0 then yield Events.Added { items = adds }
             if removes.Length <> 0 then yield Events.Deleted { items = removes } ]
 
-type Service internal (stream : Equinox.Decider<Events.Event, Fold.State>) =
+type Service internal (decider : Equinox.Decider<Events.Event, Fold.State>) =
 
     member __.Add(add : string seq, remove : string seq) : Async<int*int> =
-        stream.Transact(interpret add remove)
+        decider.Transact(interpret add remove)
 
     member __.Read() : Async<Set<string>> =
-        stream.Query id
+        decider.Query id
 
 let create resolve setId =
     let streamName = streamName setId